# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Multi-User Blog Platform API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Author {
  id: number;
  name: string;
  email: string;
  /** @nullable */
  bio?: string | null;
  /** @nullable */
  avatarUrl?: string | null;
  postCount?: number;
  createdAt: string;
}

export interface AuthorInput {
  /** @minLength 1 */
  name: string;
  email: string;
  bio?: string;
  avatarUrl?: string;
}

export interface AuthorUpdate {
  /** @minLength 1 */
  name?: string;
  bio?: string;
  avatarUrl?: string;
}

export interface Category {
  id: number;
  name: string;
  slug: string;
  /** @nullable */
  description?: string | null;
  postCount?: number;
}

export interface CategoryInput {
  /** @minLength 1 */
  name: string;
  slug: string;
  description?: string;
}

export interface Post {
  id: number;
  title: string;
  slug: string;
  /** @nullable */
  excerpt?: string | null;
  content: string;
  /** @nullable */
  coverImageUrl?: string | null;
  authorId: number;
  authorName: string;
  /** @nullable */
  authorAvatarUrl?: string | null;
  /** @nullable */
  categoryId?: number | null;
  /** @nullable */
  categoryName?: string | null;
  published: boolean;
  readingTimeMinutes?: number;
  commentCount?: number;
  createdAt: string;
  updatedAt: string;
}

export interface PostInput {
  /** @minLength 1 */
  title: string;
  slug: string;
  excerpt?: string;
  content: string;
  coverImageUrl?: string;
  authorId: number;
  categoryId?: number;
  published?: boolean;
}

export interface PostUpdate {
  /** @minLength 1 */
  title?: string;
  slug?: string;
  excerpt?: string;
  content?: string;
  coverImageUrl?: string;
  /** @nullable */
  categoryId?: number | null;
  published?: boolean;
}

export interface Comment {
  id: number;
  postId: number;
  authorName: string;
  content: string;
  createdAt: string;
}

export interface CommentInput {
  /** @minLength 1 */
  authorName: string;
  /** @minLength 1 */
  content: string;
}

export interface BlogStats {
  totalPosts: number;
  publishedPosts: number;
  totalAuthors: number;
  totalCategories: number;
  totalComments: number;
}

export type ListPostsParams = {
/**
 * @nullable
 */
categoryId?: number | null;
/**
 * @nullable
 */
authorId?: number | null;
/**
 * @nullable
 */
search?: string | null;
/**
 * @nullable
 */
limit?: number | null;
/**
 * @nullable
 */
offset?: number | null;
};