# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Multi-User Blog Platform API
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: posts
    description: Blog post operations
  - name: authors
    description: Author operations
  - name: categories
    description: Category operations
  - name: comments
    description: Comment operations

paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"

  /posts:
    get:
      operationId: listPosts
      tags: [posts]
      summary: List all published posts
      parameters:
        - name: categoryId
          in: query
          schema:
            type: ["integer", "null"]
        - name: authorId
          in: query
          schema:
            type: ["integer", "null"]
        - name: search
          in: query
          schema:
            type: ["string", "null"]
        - name: limit
          in: query
          schema:
            type: ["integer", "null"]
        - name: offset
          in: query
          schema:
            type: ["integer", "null"]
      responses:
        "200":
          description: List of posts
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Post"
    post:
      operationId: createPost
      tags: [posts]
      summary: Create a new post
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PostInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Post"

  /posts/{id}:
    get:
      operationId: getPost
      tags: [posts]
      summary: Get a single post
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Post
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Post"
        "404":
          description: Not found
    patch:
      operationId: updatePost
      tags: [posts]
      summary: Update a post
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PostUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Post"
    delete:
      operationId: deletePost
      tags: [posts]
      summary: Delete a post
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  /posts/stats/summary:
    get:
      operationId: getPostStats
      tags: [posts]
      summary: Get blog stats summary
      responses:
        "200":
          description: Stats
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BlogStats"

  /posts/recent:
    get:
      operationId: getRecentPosts
      tags: [posts]
      summary: Get recent posts
      responses:
        "200":
          description: Recent posts
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Post"

  /authors:
    get:
      operationId: listAuthors
      tags: [authors]
      summary: List all authors
      responses:
        "200":
          description: List of authors
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Author"
    post:
      operationId: createAuthor
      tags: [authors]
      summary: Create a new author
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AuthorInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Author"

  /authors/{id}:
    get:
      operationId: getAuthor
      tags: [authors]
      summary: Get an author profile
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Author
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Author"
        "404":
          description: Not found
    patch:
      operationId: updateAuthor
      tags: [authors]
      summary: Update author profile
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AuthorUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Author"

  /categories:
    get:
      operationId: listCategories
      tags: [categories]
      summary: List all categories
      responses:
        "200":
          description: Categories
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Category"
    post:
      operationId: createCategory
      tags: [categories]
      summary: Create a category
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CategoryInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Category"

  /posts/{postId}/comments:
    get:
      operationId: listComments
      tags: [comments]
      summary: List comments for a post
      parameters:
        - name: postId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Comments
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Comment"
    post:
      operationId: createComment
      tags: [comments]
      summary: Add a comment to a post
      parameters:
        - name: postId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CommentInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Comment"

  /comments/{id}:
    delete:
      operationId: deleteComment
      tags: [comments]
      summary: Delete a comment
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status

    Author:
      type: object
      required: [id, name, email, createdAt]
      properties:
        id:
          type: integer
        name:
          type: string
        email:
          type: string
        bio:
          type: ["string", "null"]
        avatarUrl:
          type: ["string", "null"]
        postCount:
          type: integer
        createdAt:
          type: string

    AuthorInput:
      type: object
      required: [name, email]
      properties:
        name:
          type: string
          minLength: 1
        email:
          type: string
        bio:
          type: string
        avatarUrl:
          type: string

    AuthorUpdate:
      type: object
      properties:
        name:
          type: string
          minLength: 1
        bio:
          type: string
        avatarUrl:
          type: string

    Category:
      type: object
      required: [id, name, slug]
      properties:
        id:
          type: integer
        name:
          type: string
        slug:
          type: string
        description:
          type: ["string", "null"]
        postCount:
          type: integer

    CategoryInput:
      type: object
      required: [name, slug]
      properties:
        name:
          type: string
          minLength: 1
        slug:
          type: string
        description:
          type: string

    Post:
      type: object
      required: [id, title, slug, content, authorId, authorName, published, createdAt, updatedAt]
      properties:
        id:
          type: integer
        title:
          type: string
        slug:
          type: string
        excerpt:
          type: ["string", "null"]
        content:
          type: string
        coverImageUrl:
          type: ["string", "null"]
        authorId:
          type: integer
        authorName:
          type: string
        authorAvatarUrl:
          type: ["string", "null"]
        categoryId:
          type: ["integer", "null"]
        categoryName:
          type: ["string", "null"]
        published:
          type: boolean
        readingTimeMinutes:
          type: integer
        commentCount:
          type: integer
        createdAt:
          type: string
        updatedAt:
          type: string

    PostInput:
      type: object
      required: [title, slug, content, authorId]
      properties:
        title:
          type: string
          minLength: 1
        slug:
          type: string
        excerpt:
          type: string
        content:
          type: string
        coverImageUrl:
          type: string
        authorId:
          type: integer
        categoryId:
          type: integer
        published:
          type: boolean

    PostUpdate:
      type: object
      properties:
        title:
          type: string
          minLength: 1
        slug:
          type: string
        excerpt:
          type: string
        content:
          type: string
        coverImageUrl:
          type: string
        categoryId:
          type: ["integer", "null"]
        published:
          type: boolean

    Comment:
      type: object
      required: [id, postId, authorName, content, createdAt]
      properties:
        id:
          type: integer
        postId:
          type: integer
        authorName:
          type: string
        content:
          type: string
        createdAt:
          type: string

    CommentInput:
      type: object
      required: [authorName, content]
      properties:
        authorName:
          type: string
          minLength: 1
        content:
          type: string
          minLength: 1

    BlogStats:
      type: object
      required: [totalPosts, totalAuthors, totalCategories, totalComments, publishedPosts]
      properties:
        totalPosts:
          type: integer
        publishedPosts:
          type: integer
        totalAuthors:
          type: integer
        totalCategories:
          type: integer
        totalComments:
          type: integer