# lib/api-zod/src/generated/types/author.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Multi-User Blog Platform API
 * OpenAPI spec version: 0.1.0
 */

export interface Author {
  id: number;
  name: string;
  email: string;
  /** @nullable */
  bio?: string | null;
  /** @nullable */
  avatarUrl?: string | null;
  postCount?: number;
  createdAt: string;
}