# lib/api-zod/src/generated/types/authorInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Multi-User Blog Platform API
 * OpenAPI spec version: 0.1.0
 */

export interface AuthorInput {
  /** @minLength 1 */
  name: string;
  email: string;
  bio?: string;
  avatarUrl?: string;
}