# lib/api-zod/src/generated/types/authorUpdate.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Multi-User Blog Platform API
 * OpenAPI spec version: 0.1.0
 */

export interface AuthorUpdate {
  /** @minLength 1 */
  name?: string;
  bio?: string;
  avatarUrl?: string;
}