# lib/api-zod/src/generated/types/post.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Multi-User Blog Platform API
 * OpenAPI spec version: 0.1.0
 */

export interface Post {
  id: number;
  title: string;
  slug: string;
  /** @nullable */
  excerpt?: string | null;
  content: string;
  /** @nullable */
  coverImageUrl?: string | null;
  authorId: number;
  authorName: string;
  /** @nullable */
  authorAvatarUrl?: string | null;
  /** @nullable */
  categoryId?: number | null;
  /** @nullable */
  categoryName?: string | null;
  published: boolean;
  readingTimeMinutes?: number;
  commentCount?: number;
  createdAt: string;
  updatedAt: string;
}