# lib/api-zod/src/generated/types/postInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Multi-User Blog Platform API
 * OpenAPI spec version: 0.1.0
 */

export interface PostInput {
  /** @minLength 1 */
  title: string;
  slug: string;
  excerpt?: string;
  content: string;
  coverImageUrl?: string;
  authorId: number;
  categoryId?: number;
  published?: boolean;
}