# lib/api-zod/src/generated/types/postUpdate.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Multi-User Blog Platform API
 * OpenAPI spec version: 0.1.0
 */

export interface PostUpdate {
  /** @minLength 1 */
  title?: string;
  slug?: string;
  excerpt?: string;
  content?: string;
  coverImageUrl?: string;
  /** @nullable */
  categoryId?: number | null;
  published?: boolean;
}